Implement the MVC architecture.
