+++ YoutubeProject/AppDelegate.swift
@@ -1,0 +1,17 @@
+//
+//  AppDelegate.swift
+//  YoutubeProject
+//
+//  Created by Felipe Lima on 26/05/23.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        return true
+    }
+}
