+++ YoutubeProject/SceneDelegate.swift
@@ -1,0 +1,19 @@
+//
+//  SceneDelegate.swift
+//  YoutubeProject
+//
+//  Created by Felipe Lima on 26/05/23.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+       
+        guard let _ = (scene as? UIWindowScene) else { return }
+    }
+}
